Fixed counter, setup LQR, changed disarm
The counter behavior was changed so that we can decrease and increase the frequency of servo publishing.

The LQR is now setup for an 8 flap vehicle, and the default gains and trim flap values are setup for 30 degrees, 6300 RPM and 19 degrees rudder.

Setup the system so that disarm just turns of the motor but keeps the servo commands.