Added DI, mode 26 is rate control, mode 29 is att control.
We changed the functions in CEMAV such that mode 26 is rate control. If RAT_CTRL is set to 1, then we will use PID. If RAT_CTRL is set to 2, then we will use state feedback. If RAT_CTRL is set to 3, then we will use NDI. Mode 29 is att control. If AT_CTRL is set to 0, then we perform PID on the angles ONLY. If AT_CTRL is set to 1, then there is an inner PID controller on rate. If AT_CTRL is set to 2, then there is an inner statefeedback controller on rate. If AT_CTRL is set to 3, then there is an inner NDI controller on rate.